Moved all the classes to their ouwn files.
