+++ pong/gameplay.py
@@ -1,5 +1,8 @@
 import pygame
 from constance import *
+from ball import *
+from paddle import *
+from player import *
 
 #Initializing pygame and puting window roperties.
 pygame.init()
@@ -8,83 +11,6 @@ pygame.display.set_caption("PongAI Master")
 left_player_score = 0
 right_player_score = 0
 
-
-#player properties class
-class Player:
-    FONTSIZE = 25
-    COLOR = WHITE
-    def __init__(self, name, score, left):
-        self.name = name
-        self.score = score
-        self.left = left
-
-    def draw(self, win):
-        font = pygame.font.SysFont(None, self.FONTSIZE)
-        score_text = font.render(f"{self.name} : {str(self.score)}", True, self.COLOR)
-
-        if self.left:
-            score_rect = score_text.get_rect(topleft=(PADDING, PADDING))
-        else:
-            score_rect = score_text.get_rect(topright=(WIDTH-PADDING, PADDING))
-
-        win.blit(score_text,score_rect)
-
-    def change_score(self, add = True):
-        if add:
-            self.score += 1
-        else:
-            self.score -= 1
-
-    def get_score(self):
-        return self.score
-    
-    def get_name(self):
-        return self.name
-
-
-
-class Paddle:
-    COLOR = WHITE
-    HEIGHT = HEIGHT
-    VELOCITY = 7
-    def __init__(self, x, y, width, height):
-        self.x = x
-        self.y = y
-        self.width = width
-        self.height = height
-
-    def draw(self, win):
-        pygame.draw.rect(win, self.COLOR,(self.x, self.y, self.width, self.height))
-
-    def move(self, up=True):
-        if up and self.y >= 0 :
-            self.y -= self.VELOCITY
-        else:
-            if self.y < HEIGHT-self.height:
-                self.y += self.VELOCITY
-
-class Ball:
-    MAX_VELOCITY = 5
-    COLOR = WHITE
-    def __init__(self, x, y, radius):
-        self.x = x
-        self.y = y
-        self.radius  = radius
-
-        self.x_velocity = self.MAX_VELOCITY
-        self.y_velocity = 0
-
-    def draw(self, win):
-        pygame.draw.circle(win, self.COLOR,(self.x,self.y),self.radius)
-
-    def move(self):
-        self.x += self.x_velocity
-        self.y += self.y_velocity
-        
-        if self.y_velocity > self.MAX_VELOCITY:
-            self.y_velocity = self.MAX_VELOCITY
-        elif self.y_velocity < -self.MAX_VELOCITY:
-            self.y_velocity = -self.MAX_VELOCITY
 
 
 #this is for moving the paddles
@@ -104,13 +30,13 @@ def handle_paddle_movement(keys, left_paddle, right_paddle):
 
 #this is for collision 
 def handle_collision(ball, left_paddle, right_paddle, left_player, right_player):
-    if (ball.y + ball.radius >= HEIGHT) or (ball.y - ball.radius <= 0):
+    if (ball.y + ball.RADIUS >= HEIGHT) or (ball.y - ball.RADIUS <= 0):
         ball.y_velocity *= -1
 
     #check if the ball hit the paddles.
     if ball.x_velocity < 0:
         if ball.y >= left_paddle.y and ball.y <= left_paddle.y + left_paddle.height:
-            if ball.x - ball.radius <= left_paddle.x + left_paddle.width:
+            if ball.x - ball.RADIUS <= left_paddle.x + left_paddle.width:
                 ball.x_velocity *= -1
                 # Change the vertical velocity based on the hit position
                 ''' relative_hit_pos = (ball.y + ball.radius) - (left_paddle.y + left_paddle.height // 2)
@@ -125,7 +51,7 @@ def handle_collision(ball, left_paddle, right_paddle, left_player, right_player)
         
     else:
         if ball.y >= right_paddle.y and ball.y <= right_paddle.y + right_paddle.height:
-            if ball.x + ball.radius >= right_paddle.x:
+            if ball.x + ball.RADIUS >= right_paddle.x:
                 ball.x_velocity *= -1
                 # Change the vertical velocity based on the hit position
                 ''' relative_hit_pos = (ball.y + ball.radius) - (right_paddle.y + right_paddle.height // 2)
@@ -139,19 +65,17 @@ def handle_collision(ball, left_paddle, right_paddle, left_player, right_player)
 
         
      # Check if the ball moves past the left or right edges
-    if ball.x - ball.radius < 0:
+    if ball.x - ball.RADIUS < 0:
         right_player.change_score()
-        ball.x = WIDTH // 2
-        ball.y = HEIGHT // 2
-        ball.x_velocity = ball.MAX_VELOCITY
-        ball.y_velocity = 0
+        ball.reset()
+        left_paddle.reset()
+        right_paddle.reset()
 
-    if ball.x + ball.radius > WIDTH:
+    if ball.x + ball.RADIUS > WIDTH:
         left_player.change_score()
-        ball.x = WIDTH // 2
-        ball.y = HEIGHT // 2
-        ball.x_velocity = -ball.MAX_VELOCITY
-        ball.y_velocity = 0 
+        ball.reset()
+        left_paddle.reset()
+        right_paddle.reset()
 
 
 #this function draws the current state of the window. 
@@ -201,7 +125,7 @@ def main():
     left_paddle = Paddle(PADDING, HEIGHT//2 - PADDLE_HEIGHT//2, PADDLE_WIDTH, PADDLE_HEIGHT)
     right_paddle = Paddle(WIDTH-PADDING-PADDLE_WIDTH, HEIGHT//2 - PADDLE_HEIGHT//2, PADDLE_WIDTH,PADDLE_HEIGHT)
 
-    ball = Ball(WIDTH//2, HEIGHT//2,BALL_RADIUS)
+    ball = Ball(WIDTH//2, HEIGHT//2)
 
     left_player = Player("Player 1", 0, True)
     right_player = Player("Player 2", 0, False)
@@ -223,10 +147,10 @@ def main():
                 run = False
                     
         if end_game(WINDOW ,left_player, right_player):
-            left_player.score = 0
-            right_player.score = 0
-            left_paddle.y = HEIGHT//2 - PADDLE_HEIGHT//2
-            right_paddle.y = HEIGHT//2 - PADDLE_HEIGHT//2
+            left_player.reset()
+            right_player.reset()
+            left_paddle.reset()
+            right_paddle.reset()
