+++ requirements.txt
@@ -1,3 +1,2 @@
-Python==3.10.4
+neat-python==0.92
 pygame==2.5.0
-
